+++ src/accountManagement/types.ts
@@ -88,10 +88,6 @@ export interface UpdateAccountObject {
      * Email address of the sales contact.
      */
     salesContactEmail?: string;
-    /**
-     * Account specific custom properties.
-     */
-    customProperties?: { [key: string]: { [key: string]: string; }; };
 }
 
 export type ApiKeyStatusEnum = "ACTIVE" | "INACTIVE";
@@ -163,10 +159,6 @@ export interface UserObject {
      * Groups
      */
     groups?: Array<string>;
-    /**
-     * User's account specific custom properties.
-     */
-    customProperties?: { [key: string]: string; };
 }
 
 export interface AddUserObject extends UserObject {