+++ types/bootstrap/models/preSharedKey.d.ts
@@ -3,6 +3,10 @@ import { BootstrapApi } from "../bootstrapApi";
 import { CallbackFn } from "../../common/interfaces";
 export declare class PreSharedKey {
     private readonly _api;
+    /**
+     * Creation time.
+     */
+    readonly createdAt?: Date;
     constructor(init: Partial<PreSharedKey>, _api: BootstrapApi);
     /**
      * Delete this PSK.